requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asyncmy==0.2.9
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
databases==0.9.0
fastapi==0.114.0
h11==0.14.0
idna==3.8
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
PyMySQL==1.1.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.4
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
